@@ -1,17 +1,7 @@
-import Vue from 'vue'
+import './assets/main.css'
+
+import { createApp } from 'vue'
 import App from './App.vue'
 
-Vue.config.productionTip = false
-
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
-
-
-function f(){
-  i = 1;         // js S2703
-
-  for (j = 0; j < array.length; j++) {  // js S2703
-    // ...
-  }
-}
+// FIXME: S1134
+createApp(App).mount('#app')
